refactored all job logic to `MainPage`
