+++ ui/src/components/JobsList.jsx
@@ -26,7 +26,7 @@ const JobsList = ({ jobs, onAddNewJob, onDelete }) => {
             <Typography variant="h4" sx={{margin: '30px'}}>All Cron Jobs</Typography>
           </Grid>
           <Grid item xs={3}>
-            <Link to="/add">
+            <Link to="/jobs/add">
               <Button sx={{ fontSize: '18px', margin: '30px' }} variant='contained' onClick={onAddNewJob}>Add New
               </Button>
             </Link>
